+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Added
 
-- Add `Flop.reset_filters/1`.
+- Add `Flop.reset_filters/1`, `Flop.to_next_page/2` and
+  `Flop.to_previous_page/1`.
 
 ### Removed